requirements.txt
asgiref==3.7.2
Django==4.2.4
django-crispy-forms==2.0
Pillow==10.0.0
psycopg2==2.9.7
python-decouple==3.8    
python-dotenv==1.0.0    
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3